requirements.txt
prettytable==3.3.0
wcwidth==0.2.5
pyyaml==6.0.1
